pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry-core.masonry.api"

[tool.poetry]
name = "ynab-api-import"
version = "0.0.0"
authors = ["Daniel Basta <ynab@basta.info>"]
description = "Library to import bank transactions via API into You Need A Budget (YNAB)"
readme = "README.md"
license = 'MIT'
packages = [{include = 'ynabapiimport'}]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = '>=3.8,<4.0'
pyyaml = '>=6.0'
requests = '>=2.28'
nordigen = '>=1.3'

[poetry.urls]
"Homepage" = "https://github.com/dnbasta/ynab-api-import"
"Bug Tracker" = "https://github.com/dnbasta/ynab-api-import/issues"
